+++ lib/dynamic_form.dart
@@ -1,6 +1,7 @@
 library dynamic_form;
 
 export 'src/simple_dynamic_form.dart';
+export 'src/login_form.dart';
 export 'src/element.dart';
 export 'src/group_elements.dart';
 export 'src/decoration_element.dart';